Moved Withdrawal struct to library
