+++ packages/protocol/src/index.ts
@@ -60,7 +60,6 @@ export * from "./settlement/messages/OutgoingMessageArgument";
 export * from "./settlement/messages/OutgoingMessage";
 export * from "./settlement/modules/NetworkStateSettlementModule";
 export * from "./settlement/messages/Deposit";
-export * from "./settlement/messages/Withdrawal";
 export { constants as ProtocolConstants } from "./Constants";
 export * from "./hashing/protokit-prefixes";
 export * from "./hashing/mina-prefixes";
