--- packages/protocol/src/settlement/messages/Withdrawal.ts
@@ -1,17 +1,0 @@
-import { Field, PublicKey, Struct, UInt64 } from "o1js";
-import { EMPTY_PUBLICKEY } from "@proto-kit/common";
-
-// TODO Remove
-export class Withdrawal extends Struct({
-  tokenId: Field,
-  address: PublicKey,
-  amount: UInt64,
-}) {
-  public static dummy() {
-    return new Withdrawal({
-      tokenId: Field(0),
-      address: EMPTY_PUBLICKEY,
-      amount: UInt64.from(0),
-    });
-  }
-}
